e2e/Cloudrun/requirements.txt
Flask==3.0.2
gunicorn==21.2.0
google-cloud
google-cloud-storage
google-cloud-bigquery